+++ packages/editor/editor-core/examples/2-comment-jira-bento.tsx
@@ -5,8 +5,8 @@ import WithEditorActions from './../src/ui/WithEditorActions';
 import CollapsedEditor from '../src/ui/CollapsedEditor';
 import ToolbarHelp from '../src/ui/ToolbarHelp';
 import { mention, emoji } from '@atlaskit/util-data-test';
+import { EmojiProvider } from '@atlaskit/emoji';
 import { mediaProvider } from './5-full-page';
-import { EmojiProvider } from '../../../elements/emoji';
 import { MockActivityResource } from '@atlaskit/activity/dist/es5/support';
 
 export type Props = {