+++ packages/media/media-client/examples/3-collection-items.tsx
@@ -2,10 +2,10 @@ import * as React from 'react';
 import { Component } from 'react';
 import { createUserMediaClient } from '@atlaskit/media-test-helpers';
 import { Subscription } from 'rxjs/Subscription';
-import { FileIdentifier } from '..';
 import { Card } from '@atlaskit/media-card';
 import Button from '@atlaskit/button';
 import { CardsWrapper, Header } from '../example-helpers/styled';
+import { FileIdentifier } from '../src';
 
 const mediaClient = createUserMediaClient();