+++ packages/navigation/atlassian-switcher/src/__tests__/unit/component-analytics.tsx
@@ -8,7 +8,7 @@ import messages from '../../utils/messages';
 import { IntlProvider } from 'react-intl';
 import createStream, { Stream } from '../../../test-helpers/stream';
 import { AnalyticsListener, UIAnalyticsEvent } from '@atlaskit/analytics-next';
-import { ProductTopItemVariation, WorklensProductType } from '../../types';
+import { WorklensProductType } from '../../types';
 
 const DefaultAtlassianSwitcher = (props: any = {}) => {
   const stubIcon = () => <span />;
@@ -87,7 +87,6 @@ const DefaultAtlassianSwitcher = (props: any = {}) => {
           messages={messages}
           hasLoaded
           hasLoadedCritical
-          productTopItemVariation={ProductTopItemVariation.mostFrequentSite}
           {...switcherLinks}
           {...props}
         />
@@ -127,7 +126,6 @@ describe('Atlassian Switcher - Component Analytics', () => {
         licensedProducts: ['jira'],
         adminLinks: ['discoverMore'],
         fixedLinks: ['people'],
-        productTopItemVariation: ProductTopItemVariation.mostFrequentSite,
         numberOfSites: 2,
       },
     });
@@ -160,7 +158,6 @@ describe('Atlassian Switcher - Component Analytics', () => {
     wrapper = mount(
       <DefaultAtlassianSwitcher
         onEventFired={eventStream}
-        productTopItemVariation={ProductTopItemVariation.currentSite}
         overrideSwitcherLinks={overrideSwitcherLinks}
       />,
     );
@@ -176,7 +173,6 @@ describe('Atlassian Switcher - Component Analytics', () => {
         licensedProducts: ['jira'],
         adminLinks: ['discoverMore'],
         fixedLinks: ['people'],
-        productTopItemVariation: ProductTopItemVariation.currentSite,
         numberOfSites: 1,
       },
     });