+++ packages/search/global-search/src/components/common/SearchResults.tsx
@@ -11,7 +11,7 @@ import { ResultsGroup } from '../../model/Result';
 import SearchError from '../SearchError';
 import deepEqual from 'deep-equal';
 import { Scope } from '../../api/types';
-import { CancelableEvent } from '../../../../quick-search';
+import { CancelableEvent } from '@atlaskit/quick-search';
 import { FilterWithMetadata } from '../../api/CrossProductSearchClient';
 
 export interface Props {