+++ packages/media/media-filmstrip/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/media-filmstrip
 
+## 34.3.8
+
+- Updated dependencies [97bab7fd28](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/97bab7fd28):
+  - @atlaskit/button@13.3.1
+  - @atlaskit/media-card@65.2.1
+  - @atlaskit/checkbox@10.0.0
+  - @atlaskit/docs@8.1.7
+
 ## 34.3.7
 
 ### Patch Changes