+++ packages/trello/textfield/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @nachos/textfield
 
+## 0.1.8
+
+### Patch Changes
+
+- [patch][9c45dd9e94](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/9c45dd9e94):
+
+  Fix main entry point not containing all exports
+
 ## 0.1.7
 
 ### Patch Changes